backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pypdf
httpx
python-dotenv
tabulate
pandas
matplotlib
seaborn
torch --index-url https://download.pytorch.org/whl/cpu
transformers
accelerate
gunicorn
